requirements.txt
# Core async framework
aiohttp>=3.9.0
aiosqlite>=0.20.0
websockets>=12.0

# Kalshi API auth (RSA-PSS signing)
cryptography>=42.0.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: for trained models (uncomment when ready)
# scikit-learn>=1.4.0
# numpy>=1.26.0
# pandas>=2.2.0
